First Prototypes
This commit involves a copy of Tanks! (file nam: tanks2) in which I begin to develop my first prototype related to Camera and Lights. The first one is inspired by Gaspar Noe's movie Enter the Void, in which the camera moves through walls and ceilings to reach the next scenes.