Second Prototypes
I added two new prototypes to the project. Both explore what's inside a tank.

Traffic:
You are a camera located on a tanks highway but you are stuck in traffic. Your ability to pierce the most solid carcass with your sight let you see inside the tanks that are moving slowly towards an end. This prototype strives to discover the humanity between the metal frame of the destroying machine. I could be push further by adding conversation between the tank’s pilots.

Volte-Face:
You are the camera and you zoom inside the tank’s barrel. You discover the inner side of the tank, with all its vulnerability. There is a close relationship between the camera and the gun since they have shared similar features in their early development. This affiliation explains the use of the expression “shooting” to describe the operation of both tools. So what happens if the camera is shooting the tank’s turret? Is it history revolving? This prototype is an infinite zoom within the tank past, present and future.