New attempt toward final prototype and journal entry
eeeeh booooy... i'm still in a very dense forest.. don't see much light... eeeeeehhhh....eeerrrrrrrggggg

New attempt:

Mix between offset in stereo and work on rhythm...

Inspired by "Einstein on the Beach", it could be a race between the left and the right stereo output in the Philip Glass way of compositing, using simple rules to create musical sentences evolving in length. Inspired by the Knee Plays in "Einstein by the Beach". The player would hit the element along the way and we would have a glimpse of description.

Maybe I'm going toward a TANKS! opera.. TANKS! in a choir singing to the god of Unity.