Final prototype: the discussion
This prototype exhibits a first-person perspective of a discussion between four creatures composed of different tanks' parts. One of them leads the conversation, talking about identity, real and virtual. Here the 3D objects speak up about their situation. This game is for them to express themselves, not for the player. The player is part of the discussion, certainly, but does s/he really have autonomy in this world praising freedom?